+++ forwarder/update_handlers/ca_update_handler.py
@@ -10,12 +10,14 @@ from forwarder.epics_to_serialisable_types import (
     ca_alarm_status_to_f142,
 )
 from caproto.threading.client import Context as CAContext
+from caproto.threading.client import PV
 from typing import Optional, Tuple, Any
 from forwarder.update_handlers.schema_publishers import schema_publishers
-
-
-def _seconds_to_nanoseconds(time_seconds: float) -> int:
-    return int(time_seconds * 1_000_000_000)
+import time
+from forwarder.kafka.kafka_helpers import (
+    publish_connection_status_message,
+    seconds_to_nanoseconds,
+)
 
 
 class CAUpdateHandler:
@@ -37,7 +39,9 @@ class CAUpdateHandler:
         self._logger = get_logger()
         self._producer = producer
         self._output_topic = output_topic
-        (self._pv,) = context.get_pvs(pv_name)
+        (self._pv,) = context.get_pvs(
+            pv_name, connection_state_callback=self._connection_state_callback
+        )
         # Subscribe with "data_type='time'" to get timestamp and alarm fields
         sub = self._pv.subscribe(data_type="time")
         sub.add_callback(self._monitor_callback)
@@ -74,7 +78,7 @@ class CAUpdateHandler:
                 return
 
         with self._cache_lock:
-            timestamp = _seconds_to_nanoseconds(response.metadata.timestamp)
+            timestamp = seconds_to_nanoseconds(response.metadata.timestamp)
             # If this is the first update or the alarm status has changed, then
             # include alarm status in message
             if (
@@ -100,6 +104,15 @@ class CAUpdateHandler:
                     timestamp,
                 )
             self._cached_update = (response, timestamp)
+
+    def _connection_state_callback(self, pv: PV, state: str):
+        publish_connection_status_message(
+            self._producer,
+            self._output_topic,
+            self._pv.name,
+            seconds_to_nanoseconds(time.time()),
+            state,
+        )
 
     def _try_to_determine_type(self, response: ReadNotifyResponse) -> bool:
         """