+++ manual_testing/add_pva_config.py
@@ -1,4 +1,11 @@
 from forwarder.kafka.kafka_helpers import create_producer
+from streaming_data_types.forwarder_config_update_rf5k import serialise_rf5k, StreamInfo
+from streaming_data_types.fbschemas.forwarder_config_update_rf5k.UpdateType import (
+    UpdateType,
+)
+from streaming_data_types.fbschemas.forwarder_config_update_rf5k.Protocol import (
+    Protocol,
+)
 import time
 
 """
@@ -7,26 +14,14 @@ import time
 
 
 if __name__ == "__main__":
-    config_message = (
-        "{"
-        '  "cmd": "add",'
-        '  "streams": ['
-        "    {"
-        '      "channel": "SIMPLE:DOUBLE3",'
-        '      "channel_provider_type": "pva",'
-        '      "converter": {'
-        '        "schema": "f142",'
-        '        "topic": "localhost:9092/pva_data"'
-        "      }"
-        "    }"
-        "  ]"
-        "}"
-    )
-
-    print(config_message)
-
     producer = create_producer("localhost:9092")
+
     producer.produce(
-        "forwarder_config", config_message.encode("utf8"), int(time.time() * 1000)
+        "forwarder_config",
+        serialise_rf5k(
+            UpdateType.ADD,
+            [StreamInfo("SIMPLE:DOUBLE3", "f142", "pva_data", Protocol.PVA,)],
+        ),
+        int(time.time() * 1000),
     )
     producer.close()