+++ tests/test_helpers/p4p_fakes.py
@@ -1,4 +1,4 @@
-from typing import Callable
+from typing import Callable, Union
 from p4p import Value
 
 
@@ -16,14 +16,16 @@ class FakeContext:
 
         self.callback: Callable = callback_undefined
 
-    def call_monitor_callback_with_fake_pv_update(self, pv_update: Value):
+    def call_monitor_callback_with_fake_pv_update(
+        self, pv_update: Union[Value, Exception]
+    ):
         self.callback(pv_update)
 
     def makeRequest(self, request_string: str) -> None:
         return None
 
     def monitor(
-        self, pv_name: str, callback: Callable, request: None
+        self, pv_name: str, callback: Callable, request: None, notify_disconnect: bool
     ) -> FakeSubscription:
         self.callback = callback
         return FakeSubscription()